Add support for Detour Road Events and upgrade projects to NET 9
